ZA5913 Eurobarometer Sample

ZA5913: Eurobarometer 81.2 (March 2014): Europeans in 2014, Financial and Economic Crisis, European Citizenship, and European Values

ZA5913_sample

Format

A data frame with 4123 rows in 36 variables:

doi

DIGITAL OBJECT IDENTIFIER

version

GESIS ARCHIVE VERSION & DATE

uniqid

UNIQUE RESPONDENT ID (CASEID BY TNS COUNTRY CODE)

isocntry

COUNTRY CODE - ISO 3166

p1

DATE OF INTERVIEW

p3

DURATION OF INTERVIEW (MINUTES)

p4

N OF PERSONS PRESENT DURING INTERVIEW

p5

RESPONDENT COOPERATION

nuts

REGION - NUTS CODES

d7

MARITAL STATUS

d8

AGE EDUCATION

d25

TYPE OF COMMUNITY

d60

DIFFICULTIES PAYING BILLS - LAST YEAR

qa10_3

EUROPEAN CENTRAL BANK - TRUST

qa10_2

EUROPEAN COMMISSION - TRUST

qa10_1

EUROPEAN PARLIAMENT - TRUST

qa7_4

TRUST IN INSTITUTIONS: EUROPEAN UNION

qa7_2

TRUST IN INSTITUTIONS: NATIONAL GOVERNMENT

qa7_3

TRUST IN INSTITUTIONS: NATIONAL PARLIAMENT

qa7_1

TRUST IN INSTITUTIONS: POLITICAL PARTIES

qa7_5

TRUST IN INSTITUTIONS: REG/LOC AUTHORITIES

qd3_1

IMPORTANT VALUES PERS: RULE OF LAW

qd3_2

IMPORTANT VALUES PERS: RESPECT HUMAN LIFE

qd3_3

IMPORTANT VALUES PERS: HUMAN RIGHTS

qd3_4

IMPORTANT VALUES PERS: INDIVIDUAL FREEDOM

qd3_5

IMPORTANT VALUES PERS: DEMOCRACY

qd3_6

IMPORTANT VALUES PERS: PEACE

qd3_7

IMPORTANT VALUES PERS: EQUALITY

qd3_8

IMPORTANT VALUES PERS: SOLIDARITY

qd3_9

IMPORTANT VALUES PERS: TOLERANCE

qd3_10

IMPORTANT VALUES PERS: RELIGION

qd3_11

IMPORTANT VALUES PERS: SELF-FULFILMENT

qd3_12

IMPORTANT VALUES PERS: RESPECT FOR CULTURES

qd3_13

IMPORTANT VALUES PERS: NONE (SPONT)

qd3_14

IMPORTANT VALUES PERS: DK

w1

WEIGHT RESULT FROM TARGET

w3

WEIGHT UNITED GERMANY ONLY

@source ZA5913: Eurobarometer 81.2 (March 2014) on gesis.org @seealso label_normalize for text normalization and label_suggest for canonisation. @family sample data files

Details

A reduced part of a raw GESIS file to demonstrate changes in variable labels and variable names.