Vocabulary To Rename Categorical Variables With Two Valid Categories

Later the filename may be omitted. The canonical_name is suggested for programmatic use in R files. The keywords should be consistent with the ELSST thesaurus.

categorical_variables_2

Format

A data frame with 6 rows in 9 variables:

filename

The name of the GESIS SPSS file where the variable can be found

r_name

The raw variable name in the GESIS SPSS file.

normalized_names

The variable label after normalization

canonical_names

The preferred variable label

keyword_1

If necessary, later new keyword columns can be added.

geo_qualifier

To explicitly refer to separate questionnaire items like the Turkish Cypriot Community

See also