Construct A Variable Of Type 'eurobarometer_labelled'

Construct A Variable Of Type 'eurobarometer_labelled'

eurobarometer_labelled(
  x = NULL,
  value_labels = NULL,
  na_values = NULL,
  na_range = NULL,
  labels_orig = NULL,
  var_name_orig = NULL,
  qb = NULL,
  conversion = NULL
)

Arguments

x
A vector of numeric values of class integer, numeric, haven_labelled or haven_labelled_spss.

value_labels
The labels attribute, defaults to NULL.

na_values
The na_values attribute, defaults to NULL.

na_range
The na_range attribute, defaults to NULL.

labels_orig
The labels_orig attribute which contains the original value_labels before harmonization or normalization, defaults to NULL.

var_name_orig
The var_name_orig attribute, which contains the original variable name, defaults to NULL.

qb
The qb attribute, which contains the identifier of the question block, defaults to NULL.

conversion
The conversion attribute which contains the name of the function that created the variable, defaults to NULL.

Value

A vector with metadata attributes of class eurobarometer_labelled.

Examples

eurobarometer_labelled(x = c(6,7,8,9,6,7,8),
   labels =c("HI" = 6, "HELLO" = 7,
             "HOWDY" = 8, "BYE" = 9),
   na_values = 9,
   qb = "Example Block",
   conversion = "eurobarometer_labelled"
)

#> <labelled_spss<double>[7]>
#> [1] 6 7 8 9 6 7 8
#> Missing values: 9
#> 
#> Labels:
#>  value label
#>      6    HI
#>      7 HELLO
#>      8 HOWDY
#>      9   BYE