Create A Corpus Of Survey Answer Items

Reads in all SPSS files in a directory, runs gesis_vocabulary_create as a wrapper function and binds the results into a single data.frame.

gesis_corpus_create(file_path)

Arguments

file_path
A file path to work from (will be changed)

Value

A data frame with seven variables, results binded together from gesis_vocabulary_create. The returned data frame has 8 variables:

var_name_orig

The name of the variable in the data file.

val_numeric_orig

The original SPSS numeric values, if applicable

var_label_orig

The original SPSS character value, if applicable

val_order_alpha

The rank number of the alphabetically sorted answer option.

val_order_length

The number of answer options in the questionnaire item.

val_label_normalized

Normalized version of the questionnaire item

filename

The name of the original GESIS file.

See also

Examples