Get A Metadata File

Retrieve a metadata file with from your data directory or the temporary directory of your current R session.

gesis_metadata_create(dat)

Arguments

dat
A data frame read by read_spss.

Value

A data frame with the original variable attributes and suggested conversions and changes.

Details

The structure of the returned tibble:

var_name_orig

The original variable name in SPSS

class_orig

The original variable class after importing withread_spss

var_label_orig

The original variable label in SPSS

var_label_norm

A normalized version of the variable labels

var_name_suggested

A partly canonized variable name.

factor_levels

A list of factor levels, i.e. value labels in SPSS

class_suggested

A suggested class conversion.

Examples

if (FALSE) {
##use your own file:
tmp <- haven::read_spss( file.path(gesis_dir, gesis_file) )
metadata <- gesis_metadata_create(tmp)
}