Harmonize A Question Block in Surveys Harmonize the variable names and the class of a question block.

Harmonize A Question Block in Surveys

Harmonize the variable names and the class of a question block.

harmonize_qb_vars(
  survey_list,
  metadata,
  id_vars = c("uniqid", "doi"),
  question_block = "socio-demography",
  var_name = "var_name_suggested",
  conversion = "conversion_suggested"
)

Arguments

survey_list
A list of data frames containing surveys.

metadata
A metadata table that contains var_name_orig and qb for the questions blocks, and var_name_suggested or its user defined alternative, class_suggested or its user defined alternative.

id_vars
A vector of ID variables to form a panel ID var, defaults to c("uniqid", "doi").

question_block
A user defined question block, or pre-defined question blocks: "id", "metadata", "weights", "socio-demography".

var_name
Defaults to var_name_suggested. Must be a column in metadata.

conversion
Defaults to conversion_suggested. Must be a column in metadata.

Value

A tibble containing all variables with harmonized names and classes, and a unique key column named panel_id.

See also

Examples

if (FALSE) {
harmonize_qb_vars (
   survey_list,
   metadata,
   question_block = "socio-demography",
   var_name = "var_name_suggested",
   conversion = "conversion_suggested"
   )
}