Vocabulary To Rename Categorical Variables With 2 Or 3 Valid Categories

The harmonization table of categorical value labels. Currently contains only labels with 2 or 3 categories. Missing labels are harmonized with the internal function harmonize_missing_values.

label_harmonization_table

Format

A data frame with 150 rows in 3 variables:

label_norm

Value labels created by label_normalize

value_numeric

Harmonized numeric values

label_harmonized

Harmonized labels

valid_range

Number of valid categories for the variable where this label can be found.

@seealso harmonize_value_labels label_normalize