Read A eurobarometer Package Example File This function is created to make panel data creation examples without relying on importing large SPSS files.

Read A eurobarometer Package Example File

This function is created to make panel data creation examples without relying on importing large SPSS files.

read_example_file(sample_file_name = "ZA7576_sample")

Arguments

sample_file_name
Any of ZA7576_sample, ZA6863_sample or ZA5913_sample.

Examples

read_example_file ( sample_file_name = 'ZA7576_sample' )

#> # A tibble: 6,579 x 53
#>    doi   version uniqid isocntry       p1      p2    p3      p4      p5
#>    <chr> <chr>    <dbl> <chr>    <dbl+lb> <dbl+l> <dbl> <dbl+l> <dbl+l>
#>  1 doi:~ 1.0.0 ~ 2.00e8 DE-W     12 [Tue~ 2 [8 -~    64 1 [Two~ 3 [Ave~
#>  2 doi:~ 1.0.0 ~ 2.00e8 DE-W     13 [Wed~ 3 [13 ~    78 1 [Two~ 2 [Fai~
#>  3 doi:~ 1.0.0 ~ 2.00e8 DE-W     18 [Mon~ 2 [8 -~    44 1 [Two~ 1 [Exc~
#>  4 doi:~ 1.0.0 ~ 2.00e8 DE-W     18 [Mon~ 3 [13 ~    55 1 [Two~ 1 [Exc~
#>  5 doi:~ 1.0.0 ~ 2.00e8 DE-W     19 [Tue~ 2 [8 -~    44 1 [Two~ 1 [Exc~
#>  6 doi:~ 1.0.0 ~ 2.00e8 DE-W      7 [Thu~ 2 [8 -~    41 1 [Two~ 1 [Exc~
#>  7 doi:~ 1.0.0 ~ 2.00e8 DE-W     12 [Tue~ 3 [13 ~    29 1 [Two~ 2 [Fai~
#>  8 doi:~ 1.0.0 ~ 2.00e8 DE-W      8 [Fri~ 3 [13 ~    41 1 [Two~ 2 [Fai~
#>  9 doi:~ 1.0.0 ~ 2.00e8 DE-W      7 [Thu~ 2 [8 -~    44 1 [Two~ 1 [Exc~
#> 10 doi:~ 1.0.0 ~ 2.00e8 DE-W      7 [Thu~ 2 [8 -~    44 1 [Two~ 1 [Exc~
#> # ... with 6,569 more rows, and 44 more variables: nuts <chr+lbl>,
#> #   d7 <dbl+lbl>, d8 <dbl+lbl>, d25 <dbl+lbl>, d60 <dbl+lbl>, qa14_5 <dbl+lbl>,
#> #   qa14_3 <dbl+lbl>, qa14_2 <dbl+lbl>, qa14_4 <dbl+lbl>, qa14_1 <dbl+lbl>,
#> #   qa6a_5 <dbl+lbl>, qa6a_10 <dbl+lbl>, qa6b_2 <dbl+lbl>, qa6a_3 <dbl+lbl>,
#> #   qa6a_1 <dbl+lbl>, qa6b_4 <dbl+lbl>, qa6a_8 <dbl+lbl>, qa6a_9 <dbl+lbl>,
#> #   qa6a_4 <dbl+lbl>, qa6a_2 <dbl+lbl>, qa6b_1 <dbl+lbl>, qa6a_6 <dbl+lbl>,
#> #   qa6a_7 <dbl+lbl>, qa6a_11 <dbl+lbl>, qa6b_3 <dbl+lbl>, qd6.1 <dbl+lbl>,
#> #   qd6.2 <dbl+lbl>, qd6.3 <dbl+lbl>, qd6.4 <dbl+lbl>, qd6.5 <dbl+lbl>,
#> #   qd6.6 <dbl+lbl>, qd6.7 <dbl+lbl>, qd6.8 <dbl+lbl>, qd6.9 <dbl+lbl>,
#> #   qd6.10 <dbl+lbl>, qd6.11 <dbl+lbl>, qd6.12 <dbl+lbl>, qd6.13 <dbl+lbl>,
#> #   qd6.14 <dbl+lbl>, qg1b <dbl+lbl>, qg8 <dbl+lbl>, w1 <dbl>, w3 <dbl>,
#> #   wex <dbl>