Read A Survey File

Read a survey into a survey data.frame, and record its important metadata: information about the Eurobarometer wave, the filename, the digital object identifier, the timeframe of the survey and a short description.

read_rds(file, id = NULL, doi = NULL)

read_spss(
  file,
  user_na = NULL,
  col_select = NULL,
  skip = NULL,
  n_max = NULL,
  .name_repair = "unique",
  id = NULL,
  filename = NULL,
  doi = NULL
)

Arguments

file
A re-saved survey, imported with haven::read_spss

id
An identifier of the tibble, if omitted, defaults to the file name.

doi
An optional document object identifier.

user_na
If TRUE variables with user defined missing will be read into labelled_spss() objects. If FALSE, the default, user-defined missings will be converted to NA.

col_select
One or more selection expressions, like in dplyr::select(). Use c() or list() to use more than one expression. See ?dplyr::select for details on available selection options. Only the specified columns will be read from data_file.

skip
Number of lines to skip before reading data.

n_max
Maximum number of lines to read.

.name_repair
Treatment of problematic column names:

"minimal": No name repair or checks, beyond basic existence,

"unique": Make sure names are unique and not empty,

"check_unique": (default value), no name repair, but check they are unique,

"universal": Make the names unique and syntactic

a function: apply custom name repair (e.g., .name_repair = make.names for names in the style of base R).

A purrr-style anonymous function, see rlang::as_function()

This argument is passed on as repair to vctrs::vec_as_names(). See there for more details on these terms and the strategies used to enforce them.

filename
An import file name.

Value

The survey data in a retorharmonize_survey class which records important metadata about the survey contents.

The survey data in a retorharmonize_survey class which records important metadata about the survey contents.

Details

The retorharmonize_survey class extends the metadata of the survey data frame, but otherwise it inherits all methods of a modern tibble data frame.

See also

Examples

path <- system.file("examples", "ZA7576.rds", package = "eurobarometer")
read_eb <- read_rds(path)

#> Warning: cannot open compressed file '', probable reason 'Invalid argument'

#> Error in gzfile(file, "rb"): cannot open the connection

attributes(read_eb)

#> Error in eval(expr, envir, enclos): object 'read_eb' not found

path <- system.file("examples", "iris1.sav", package = "eurobarometer")
read_spss <- read_spss(path)

#> Error in retroharmonize::read_spss(file, id = id, filename = filename,     user_na = user_na, col_select = col_select, n_max = n_max,     .name_repair = "unique", id = id, filename = filename, doi = doi): formal argument "id" matched by multiple actual arguments

attributes(read_eb)

#> Error in eval(expr, envir, enclos): object 'read_eb' not found