Read Survey Files Import surveys into a list. Add filename as a constant to each element of the list.

Read Survey Files

Import surveys into a list. Add filename as a constant to each element of the list.

read_surveys(import_file_names, .f = "read_example_file")

Arguments

import_file_names
A vector of file names to import.

.f
A function to import the surveys with. Defaults to read_example_file.

Value

A list of the surveys. Each element of the list is a data frame. The respective file names are added to each data frame as a constant column filename.

Examples

import_file_names <- c(
'ZA7576_sample','ZA7562_sample','ZA7489_sample'
)
read_surveys (import_file_names, .f = 'read_example_file' )

#> [[1]]
#> # A tibble: 6,579 x 106
#>    doi   version uniqid isocntry       p1      p2    p3      p4      p5      d7
#>    <chr> <chr>    <dbl> <chr>    <dbl+lb> <dbl+l> <dbl> <dbl+l> <dbl+l> <dbl+l>
#>  1 doi:~ 1.0.0 ~ 2.00e8 DE-W     12 [Tue~ 2 [8 -~    64 1 [Two~ 3 [Ave~ 1 [(Re~
#>  2 doi:~ 1.0.0 ~ 2.00e8 DE-W     13 [Wed~ 3 [13 ~    78 1 [Two~ 2 [Fai~ 9 [Sin~
#>  3 doi:~ 1.0.0 ~ 2.00e8 DE-W     18 [Mon~ 2 [8 -~    44 1 [Two~ 1 [Exc~ 1 [(Re~
#>  4 doi:~ 1.0.0 ~ 2.00e8 DE-W     18 [Mon~ 3 [13 ~    55 1 [Two~ 1 [Exc~ 3 [(Re~
#>  5 doi:~ 1.0.0 ~ 2.00e8 DE-W     19 [Tue~ 2 [8 -~    44 1 [Two~ 1 [Exc~ 1 [(Re~
#>  6 doi:~ 1.0.0 ~ 2.00e8 DE-W      7 [Thu~ 2 [8 -~    41 1 [Two~ 1 [Exc~ 1 [(Re~
#>  7 doi:~ 1.0.0 ~ 2.00e8 DE-W     12 [Tue~ 3 [13 ~    29 1 [Two~ 2 [Fai~ 9 [Sin~
#>  8 doi:~ 1.0.0 ~ 2.00e8 DE-W      8 [Fri~ 3 [13 ~    41 1 [Two~ 2 [Fai~ 9 [Sin~
#>  9 doi:~ 1.0.0 ~ 2.00e8 DE-W      7 [Thu~ 2 [8 -~    44 1 [Two~ 1 [Exc~ 5 [Sin~
#> 10 doi:~ 1.0.0 ~ 2.00e8 DE-W      7 [Thu~ 2 [8 -~    44 1 [Two~ 1 [Exc~ 2 [(Re~
#> # ... with 6,569 more rows, and 96 more variables: d8 <dbl+lbl>, d25 <dbl+lbl>,
#> #   d60 <dbl+lbl>, qa14_5 <dbl+lbl>, qa14_5 <dbl+lbl>, qa14_5 <dbl+lbl>,
#> #   qa14_5 <dbl+lbl>, qa14_3 <dbl+lbl>, qa14_3 <dbl+lbl>, qa14_3 <dbl+lbl>,
#> #   qa14_3 <dbl+lbl>, qa14_2 <dbl+lbl>, qa14_2 <dbl+lbl>, qa14_2 <dbl+lbl>,
#> #   qa14_2 <dbl+lbl>, qa14_4 <dbl+lbl>, qa14_4 <dbl+lbl>, qa14_4 <dbl+lbl>,
#> #   qa14_4 <dbl+lbl>, qa14_1 <dbl+lbl>, qa14_1 <dbl+lbl>, qa14_1 <dbl+lbl>,
#> #   qa14_1 <dbl+lbl>, qa6a_5 <dbl+lbl>, qa6a_5 <dbl+lbl>, qa6a_5 <dbl+lbl>,
#> #   qa6a_5 <dbl+lbl>, qa6a_10 <dbl+lbl>, qa6a_10 <dbl+lbl>, qa6a_10 <dbl+lbl>,
#> #   qa6a_10 <dbl+lbl>, qa6b_2 <dbl+lbl>, qa6b_2 <dbl+lbl>, qa6b_2 <dbl+lbl>,
#> #   qa6b_2 <dbl+lbl>, qa6a_3 <dbl+lbl>, qa6a_3 <dbl+lbl>, qa6a_3 <dbl+lbl>,
#> #   qa6a_3 <dbl+lbl>, qa6a_1 <dbl+lbl>, qa6a_1 <dbl+lbl>, qa6a_1 <dbl+lbl>,
#> #   qa6a_1 <dbl+lbl>, qa6b_4 <dbl+lbl>, qa6b_4 <dbl+lbl>, qa6b_4 <dbl+lbl>,
#> #   qa6b_4 <dbl+lbl>, qa6a_8 <dbl+lbl>, qa6a_8 <dbl+lbl>, qa6a_8 <dbl+lbl>,
#> #   qa6a_8 <dbl+lbl>, qa6a_9 <dbl+lbl>, qa6a_9 <dbl+lbl>, qa6a_9 <dbl+lbl>,
#> #   qa6a_9 <dbl+lbl>, qa6a_4 <dbl+lbl>, qa6a_4 <dbl+lbl>, qa6a_4 <dbl+lbl>,
#> #   qa6a_4 <dbl+lbl>, qa6a_2 <dbl+lbl>, qa6a_2 <dbl+lbl>, qa6a_2 <dbl+lbl>,
#> #   qa6a_2 <dbl+lbl>, qa6b_1 <dbl+lbl>, qa6b_1 <dbl+lbl>, qa6b_1 <dbl+lbl>,
#> #   qa6b_1 <dbl+lbl>, qa6a_6 <dbl+lbl>, qa6a_6 <dbl+lbl>, qa6a_6 <dbl+lbl>,
#> #   qa6a_6 <dbl+lbl>, qa6a_7 <dbl+lbl>, qa6a_7 <dbl+lbl>, qa6a_7 <dbl+lbl>,
#> #   qa6a_7 <dbl+lbl>, qa6a_11 <dbl+lbl>, qa6a_11 <dbl+lbl>, qa6a_11 <dbl+lbl>,
#> #   qa6a_11 <dbl+lbl>, qa6b_3 <dbl+lbl>, qa6b_3 <dbl+lbl>, qa6b_3 <dbl+lbl>,
#> #   qa6b_3 <dbl+lbl>, d71a_1 <dbl+lbl>, d71a_2 <dbl+lbl>, d71a_3 <dbl+lbl>,
#> #   d71b_1 <dbl+lbl>, d71b_2 <dbl+lbl>, d71b_3 <dbl+lbl>, qg1b <dbl+lbl>,
#> #   qg8 <dbl+lbl>, d77 <dbl+lbl>, w1 <dbl>, w3 <dbl>, wex <dbl>, filename <chr>
#> 
#> [[2]]
#> # A tibble: 5,074 x 51
#>    doi   version uniqid isocntry       p1      p2    p3      p4      p5       d7
#>    <chr> <chr>    <dbl> <chr>    <dbl+lb> <dbl+l> <dbl> <dbl+l> <dbl+l> <dbl+lb>
#>  1 doi:~ 1.0.0 ~ 2.00e8 DE-W      7 [Thu~ 3 [13 ~    29 1 [Two~ 2 [Fai~  1 [(Re~
#>  2 doi:~ 1.0.0 ~ 2.00e8 DE-W      4 [Mon~ 3 [13 ~    29 1 [Two~ 1 [Exc~  9 [Sin~
#>  3 doi:~ 1.0.0 ~ 2.00e8 DE-W     11 [Mon~ 2 [8 -~    24 1 [Two~ 1 [Exc~  2 [(Re~
#>  4 doi:~ 1.0.0 ~ 2.00e8 DE-W      5 [Tue~ 3 [13 ~    29 1 [Two~ 1 [Exc~  5 [Sin~
#>  5 doi:~ 1.0.0 ~ 2.00e8 DE-W      5 [Tue~ 3 [13 ~    42 1 [Two~ 1 [Exc~  2 [(Re~
#>  6 doi:~ 1.0.0 ~ 2.00e8 DE-W      5 [Tue~ 3 [13 ~    84 1 [Two~ 1 [Exc~  9 [Sin~
#>  7 doi:~ 1.0.0 ~ 2.00e8 DE-W     12 [Tue~ 2 [8 -~    34 1 [Two~ 1 [Exc~ 11 [Div~
#>  8 doi:~ 1.0.0 ~ 2.00e8 DE-W     12 [Tue~ 3 [13 ~    51 2 [Thr~ 1 [Exc~  1 [(Re~
#>  9 doi:~ 1.0.0 ~ 2.00e8 DE-W      4 [Mon~ 4 [17 ~    21 1 [Two~ 1 [Exc~  1 [(Re~
#> 10 doi:~ 1.0.0 ~ 2.00e8 DE-W     13 [Wed~ 3 [13 ~    40 1 [Two~ 1 [Exc~  1 [(Re~
#> # ... with 5,064 more rows, and 41 more variables: d8 <dbl+lbl>, d25 <dbl+lbl>,
#> #   d60 <dbl+lbl>, qa3_6 <dbl+lbl>, qa3_6 <dbl+lbl>, qa3_6 <dbl+lbl>,
#> #   qa3_1 <dbl+lbl>, qa3_1 <dbl+lbl>, qa3_1 <dbl+lbl>, qa3_4 <dbl+lbl>,
#> #   qa3_4 <dbl+lbl>, qa3_4 <dbl+lbl>, qa3_5 <dbl+lbl>, qa3_5 <dbl+lbl>,
#> #   qa3_5 <dbl+lbl>, qa3_2 <dbl+lbl>, qa3_2 <dbl+lbl>, qa3_2 <dbl+lbl>,
#> #   qa3_3 <dbl+lbl>, qa3_3 <dbl+lbl>, qa3_3 <dbl+lbl>, d71_1 <dbl+lbl>,
#> #   d71_2 <dbl+lbl>, d71_3 <dbl+lbl>, qc9.1 <dbl+lbl>, qc9.2 <dbl+lbl>,
#> #   qc9.3 <dbl+lbl>, qc9.4 <dbl+lbl>, qc9.5 <dbl+lbl>, qc9.6 <dbl+lbl>,
#> #   qc9.7 <dbl+lbl>, qc9.8 <dbl+lbl>, qc9.9 <dbl+lbl>, qc9.10 <dbl+lbl>,
#> #   qc9.11 <dbl+lbl>, qc10 <dbl+lbl>, d77 <dbl+lbl>, w1 <dbl>, w3 <dbl>,
#> #   wex <dbl>, filename <chr>
#> 
#> [[3]]
#> # A tibble: 6,613 x 134
#>    doi   version uniqid isocntry      p1      p2    p3      p4      p5       d7
#>    <chr> <chr>    <dbl> <chr>    <dbl+l> <dbl+l> <dbl> <dbl+l> <dbl+l> <dbl+lb>
#>  1 doi:~ 1.0.0 ~ 1.00e9 DE-W     2 [Sun~ 4 [17 ~    51 1 [Two~ 2 [Fai~  1 [(Re~
#>  2 doi:~ 1.0.0 ~ 1.00e9 DE-W     6 [Thu~ 4 [17 ~    34 1 [Two~ 1 [Exc~  2 [(Re~
#>  3 doi:~ 1.0.0 ~ 1.00e9 DE-W     7 [Fri~ 2 [8 -~    30 1 [Two~ 2 [Fai~  2 [(Re~
#>  4 doi:~ 1.0.0 ~ 1.00e9 DE-W     9 [Sun~ 3 [13 ~    83 1 [Two~ 1 [Exc~  2 [(Re~
#>  5 doi:~ 1.0.0 ~ 1.00e9 DE-W     9 [Sun~ 2 [8 -~    76 1 [Two~ 2 [Fai~  2 [(Re~
#>  6 doi:~ 1.0.0 ~ 1.00e9 DE-W     8 [Sat~ 2 [8 -~    47 1 [Two~ 2 [Fai~  9 [Sin~
#>  7 doi:~ 1.0.0 ~ 1.00e9 DE-W     8 [Sat~ 3 [13 ~    47 1 [Two~ 2 [Fai~ 14 [Wid~
#>  8 doi:~ 1.0.0 ~ 1.00e9 DE-W     7 [Fri~ 2 [8 -~    46 1 [Two~ 2 [Fai~ 11 [Div~
#>  9 doi:~ 1.0.0 ~ 1.00e9 DE-W     9 [Sun~ 2 [8 -~    45 1 [Two~ 2 [Fai~  2 [(Re~
#> 10 doi:~ 1.0.0 ~ 1.00e9 DE-W     6 [Thu~ 2 [8 -~    43 1 [Two~ 2 [Fai~  1 [(Re~
#> # ... with 6,603 more rows, and 124 more variables: d8 <dbl+lbl>,
#> #   d25 <dbl+lbl>, d60 <dbl+lbl>, qa13_5 <dbl+lbl>, qa13_5 <dbl+lbl>,
#> #   qa13_5 <dbl+lbl>, qa13_5 <dbl+lbl>, qa13_3 <dbl+lbl>, qa13_3 <dbl+lbl>,
#> #   qa13_3 <dbl+lbl>, qa13_3 <dbl+lbl>, qa13_2 <dbl+lbl>, qa13_2 <dbl+lbl>,
#> #   qa13_2 <dbl+lbl>, qa13_4 <dbl+lbl>, qa13_4 <dbl+lbl>, qa13_4 <dbl+lbl>,
#> #   qa13_4 <dbl+lbl>, qa13_1 <dbl+lbl>, qa13_1 <dbl+lbl>, qa13_1 <dbl+lbl>,
#> #   qa8a_9 <dbl+lbl>, qa8a_9 <dbl+lbl>, qa8a_9 <dbl+lbl>, qa8a_9 <dbl+lbl>,
#> #   qa8a_14 <dbl+lbl>, qa8a_14 <dbl+lbl>, qa8a_14 <dbl+lbl>, qa8a_14 <dbl+lbl>,
#> #   qa8b_7 <dbl+lbl>, qa8b_7 <dbl+lbl>, qa8b_7 <dbl+lbl>, qa8b_7 <dbl+lbl>,
#> #   qa8a_4 <dbl+lbl>, qa8a_4 <dbl+lbl>, qa8a_4 <dbl+lbl>, qa8a_4 <dbl+lbl>,
#> #   qa8b_4 <dbl+lbl>, qa8b_4 <dbl+lbl>, qa8b_4 <dbl+lbl>, qa8b_4 <dbl+lbl>,
#> #   qa8a_7 <dbl+lbl>, qa8a_7 <dbl+lbl>, qa8a_7 <dbl+lbl>, qa8a_7 <dbl+lbl>,
#> #   qa8a_12 <dbl+lbl>, qa8a_12 <dbl+lbl>, qa8a_12 <dbl+lbl>, qa8a_12 <dbl+lbl>,
#> #   qa8a_13 <dbl+lbl>, qa8a_13 <dbl+lbl>, qa8a_13 <dbl+lbl>, qa8a_13 <dbl+lbl>,
#> #   qa8a_5 <dbl+lbl>, qa8a_5 <dbl+lbl>, qa8a_5 <dbl+lbl>, qa8a_5 <dbl+lbl>,
#> #   qa8b_5 <dbl+lbl>, qa8b_5 <dbl+lbl>, qa8b_5 <dbl+lbl>, qa8b_5 <dbl+lbl>,
#> #   qa8a_8 <dbl+lbl>, qa8a_8 <dbl+lbl>, qa8a_8 <dbl+lbl>, qa8a_8 <dbl+lbl>,
#> #   qa8a_6 <dbl+lbl>, qa8a_6 <dbl+lbl>, qa8a_6 <dbl+lbl>, qa8a_6 <dbl+lbl>,
#> #   qa8b_6 <dbl+lbl>, qa8b_6 <dbl+lbl>, qa8b_6 <dbl+lbl>, qa8b_6 <dbl+lbl>,
#> #   qa8a_10 <dbl+lbl>, qa8a_10 <dbl+lbl>, qa8a_10 <dbl+lbl>, qa8a_10 <dbl+lbl>,
#> #   qa8a_2 <dbl+lbl>, qa8a_2 <dbl+lbl>, qa8a_2 <dbl+lbl>, qa8a_2 <dbl+lbl>,
#> #   qa8b_2 <dbl+lbl>, qa8b_2 <dbl+lbl>, qa8b_2 <dbl+lbl>, qa8b_2 <dbl+lbl>,
#> #   qa8a_11 <dbl+lbl>, qa8a_11 <dbl+lbl>, qa8a_11 <dbl+lbl>, qa8a_11 <dbl+lbl>,
#> #   qa8a_3 <dbl+lbl>, qa8a_3 <dbl+lbl>, qa8a_3 <dbl+lbl>, qa8a_3 <dbl+lbl>,
#> #   qa8b_3 <dbl+lbl>, qa8b_3 <dbl+lbl>, qa8b_3 <dbl+lbl>, qa8b_3 <dbl+lbl>,
#> #   qa8a_15 <dbl+lbl>, qa8a_15 <dbl+lbl>, qa8a_15 <dbl+lbl>, ...
#>