Read A Survey File

Read a survey into a survey data.frame, and record its important metadata: information about the Eurobarometer wave, the filename, the digital object identifier, the timeframe of the survey and a short description.

Read a survey into a survey data.frame, and record its important metadata: information about the Eurobarometer wave, the filename, the digital object identifier, the timeframe of the survey and a short description.

read_rds(file, id = NULL, doi = NULL)

read_spss(
  file,
  user_na = TRUE,
  id = NULL,
  filename = NULL,
  doi = NULL,
  .name_repair = "unique"
)

read_rds(file, id = NULL, doi = NULL)

read_spss(
  file,
  user_na = TRUE,
  id = NULL,
  filename = NULL,
  doi = NULL,
  .name_repair = "unique"
)

Arguments

file
A re-saved survey, imported with haven::read_spss

id
An identifier of the tibble, if omitted, defaults to the file name.

doi
An optional document object identifier.

filename
An import file name.

Value

The survey data in a retorharmonize_survey class which records important metadata about the survey contents.

The survey data in a retorharmonize_survey class which records important metadata about the survey contents.

The survey data in a retorharmonize_survey class which records important metadata about the survey contents.

The survey data in a retorharmonize_survey class which records important metadata about the survey contents.

Details

The retorharmonize_survey class extends the metadata of the survey data frame, but otherwise it inherits all methods of a modern tibble data frame.

The retorharmonize_survey class extends the metadata of the survey data frame, but otherwise it inherits all methods of a modern tibble data frame.

See also

Examples

path <- system.file("examples", "ZA7576.rds", package = "eurobarometer")
read_eb <- read_rds(path)

#> Survey read:

#> id: ZA7576

#> filename: ZA7576.rds

#> doi: doi:10.4232/1.13393

attributes(read_eb)

#> $names
#>  [1] "rowid"    "doi"      "version"  "uniqid"   "caseid"   "serialid"
#>  [7] "isocntry" "p1"       "p2"       "p3"       "p4"       "p5"      
#> [13] "nuts"     "d7"       "d8"       "d25"      "d60"      "qa14_5"  
#> [19] "qa14_3"   "qa14_2"   "qa14_4"   "qa14_1"   "qa6a_5"   "qa6a_10" 
#> [25] "qa6b_2"   "qa6a_3"   "qa6a_1"   "qa6b_4"   "qa6a_8"   "qa6a_9"  
#> [31] "qa6a_4"   "qa6a_2"   "qa6b_1"   "qa6a_6"   "qa6a_7"   "qa6a_11" 
#> [37] "qa6b_3"   "qd6.1"    "qd6.2"    "qd6.3"    "qd6.4"    "qd6.5"   
#> [43] "qd6.6"    "qd6.7"    "qd6.8"    "qd6.9"    "qd6.10"   "qd6.11"  
#> [49] "qd6.12"   "qd6.13"   "qd6.14"   "qg1b"     "qg8"      "w1"      
#> [55] "w3"       "wex"     
#> 
#> $row.names
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45
#> 
#> $class
#> [1] "survey"     "tbl_df"     "tbl"        "data.frame"
#> 
#> $id
#> [1] "ZA7576"
#> 
#> $filename
#> [1] "ZA7576.rds"
#> 
#> $doi
#> [1] "doi:10.4232/1.13393"
#> 
#> $gesis_file_version
#> [1] NA
#> 
#> $wave
#> [1] "Eurobarometer 91.5"
#> 
#> $timeframe
#> [1] "June-July 2019"
#> 
#> $eb_description
#> [1] "Standard Eurobarometer, European Parliament Post-election survey 2019, and EU citizens and development cooperation"
#> 

path <- system.file("examples", "iris1.sav", package = "eurobarometer")
read_spss <- read_spss(path)

#> Survey read:

#> id: iris1

#> filename: iris1.sav

attributes(read_eb)

#> $names
#>  [1] "rowid"    "doi"      "version"  "uniqid"   "caseid"   "serialid"
#>  [7] "isocntry" "p1"       "p2"       "p3"       "p4"       "p5"      
#> [13] "nuts"     "d7"       "d8"       "d25"      "d60"      "qa14_5"  
#> [19] "qa14_3"   "qa14_2"   "qa14_4"   "qa14_1"   "qa6a_5"   "qa6a_10" 
#> [25] "qa6b_2"   "qa6a_3"   "qa6a_1"   "qa6b_4"   "qa6a_8"   "qa6a_9"  
#> [31] "qa6a_4"   "qa6a_2"   "qa6b_1"   "qa6a_6"   "qa6a_7"   "qa6a_11" 
#> [37] "qa6b_3"   "qd6.1"    "qd6.2"    "qd6.3"    "qd6.4"    "qd6.5"   
#> [43] "qd6.6"    "qd6.7"    "qd6.8"    "qd6.9"    "qd6.10"   "qd6.11"  
#> [49] "qd6.12"   "qd6.13"   "qd6.14"   "qg1b"     "qg8"      "w1"      
#> [55] "w3"       "wex"     
#> 
#> $row.names
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45
#> 
#> $class
#> [1] "survey"     "tbl_df"     "tbl"        "data.frame"
#> 
#> $id
#> [1] "ZA7576"
#> 
#> $filename
#> [1] "ZA7576.rds"
#> 
#> $doi
#> [1] "doi:10.4232/1.13393"
#> 
#> $gesis_file_version
#> [1] NA
#> 
#> $wave
#> [1] "Eurobarometer 91.5"
#> 
#> $timeframe
#> [1] "June-July 2019"
#> 
#> $eb_description
#> [1] "Standard Eurobarometer, European Parliament Post-election survey 2019, and EU citizens and development cooperation"
#> 

path <- system.file("examples", "ZA7576.rds", package = "eurobarometer")
read_eb <- read_rds(path)

#> Survey read:

#> id: ZA7576

#> filename: ZA7576.rds

#> doi: doi:10.4232/1.13393

attributes(read_eb)

#> $names
#>  [1] "rowid"    "doi"      "version"  "uniqid"   "caseid"   "serialid"
#>  [7] "isocntry" "p1"       "p2"       "p3"       "p4"       "p5"      
#> [13] "nuts"     "d7"       "d8"       "d25"      "d60"      "qa14_5"  
#> [19] "qa14_3"   "qa14_2"   "qa14_4"   "qa14_1"   "qa6a_5"   "qa6a_10" 
#> [25] "qa6b_2"   "qa6a_3"   "qa6a_1"   "qa6b_4"   "qa6a_8"   "qa6a_9"  
#> [31] "qa6a_4"   "qa6a_2"   "qa6b_1"   "qa6a_6"   "qa6a_7"   "qa6a_11" 
#> [37] "qa6b_3"   "qd6.1"    "qd6.2"    "qd6.3"    "qd6.4"    "qd6.5"   
#> [43] "qd6.6"    "qd6.7"    "qd6.8"    "qd6.9"    "qd6.10"   "qd6.11"  
#> [49] "qd6.12"   "qd6.13"   "qd6.14"   "qg1b"     "qg8"      "w1"      
#> [55] "w3"       "wex"     
#> 
#> $row.names
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45
#> 
#> $class
#> [1] "survey"     "tbl_df"     "tbl"        "data.frame"
#> 
#> $id
#> [1] "ZA7576"
#> 
#> $filename
#> [1] "ZA7576.rds"
#> 
#> $doi
#> [1] "doi:10.4232/1.13393"
#> 
#> $gesis_file_version
#> [1] NA
#> 
#> $wave
#> [1] "Eurobarometer 91.5"
#> 
#> $timeframe
#> [1] "June-July 2019"
#> 
#> $eb_description
#> [1] "Standard Eurobarometer, European Parliament Post-election survey 2019, and EU citizens and development cooperation"
#> 

path <- system.file("examples", "iris1.sav", package = "eurobarometer")
read_spss <- read_spss(path)

#> Survey read:

#> id: iris1

#> filename: iris1.sav

attributes(read_eb)

#> $names
#>  [1] "rowid"    "doi"      "version"  "uniqid"   "caseid"   "serialid"
#>  [7] "isocntry" "p1"       "p2"       "p3"       "p4"       "p5"      
#> [13] "nuts"     "d7"       "d8"       "d25"      "d60"      "qa14_5"  
#> [19] "qa14_3"   "qa14_2"   "qa14_4"   "qa14_1"   "qa6a_5"   "qa6a_10" 
#> [25] "qa6b_2"   "qa6a_3"   "qa6a_1"   "qa6b_4"   "qa6a_8"   "qa6a_9"  
#> [31] "qa6a_4"   "qa6a_2"   "qa6b_1"   "qa6a_6"   "qa6a_7"   "qa6a_11" 
#> [37] "qa6b_3"   "qd6.1"    "qd6.2"    "qd6.3"    "qd6.4"    "qd6.5"   
#> [43] "qd6.6"    "qd6.7"    "qd6.8"    "qd6.9"    "qd6.10"   "qd6.11"  
#> [49] "qd6.12"   "qd6.13"   "qd6.14"   "qg1b"     "qg8"      "w1"      
#> [55] "w3"       "wex"     
#> 
#> $row.names
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45
#> 
#> $class
#> [1] "survey"     "tbl_df"     "tbl"        "data.frame"
#> 
#> $id
#> [1] "ZA7576"
#> 
#> $filename
#> [1] "ZA7576.rds"
#> 
#> $doi
#> [1] "doi:10.4232/1.13393"
#> 
#> $gesis_file_version
#> [1] NA
#> 
#> $wave
#> [1] "Eurobarometer 91.5"
#> 
#> $timeframe
#> [1] "June-July 2019"
#> 
#> $eb_description
#> [1] "Standard Eurobarometer, European Parliament Post-election survey 2019, and EU citizens and development cooperation"
#>