Read Survey Files Import surveys into a list. Adds filename as a constant to each element of the list.

Read Survey Files

Import surveys into a list. Adds filename as a constant to each element of the list.

read_surveys(import_file_names, .f = "read_example_file")

Arguments

import_file_names
A vector of file names to import.

.f
A function to import the surveys with. Defaults to 'read_example_file'. For SPSS files, 'read_spss_survey' is recommended, which is a well-parameterised version of read_spss.

Value

A list of the surveys. Each element of the list is a data frame. The respective file names are added to each data frame as a constant column filename.

Examples

import_file_names <- c(
'ZA7576_sample','ZA6863_sample','ZA5913_sample'
)
read_surveys (import_file_names, .f = 'read_example_file' )

#> [[1]]
#> # A tibble: 6,579 x 54
#>    doi   version uniqid isocntry       p1      p2    p3      p4      p5
#>    <chr> <chr>    <dbl> <chr>    <dbl+lb> <dbl+l> <dbl> <dbl+l> <dbl+l>
#>  1 doi:~ 1.0.0 ~ 2.00e8 DE-W     12 [Tue~ 2 [8 -~    64 1 [Two~ 3 [Ave~
#>  2 doi:~ 1.0.0 ~ 2.00e8 DE-W     13 [Wed~ 3 [13 ~    78 1 [Two~ 2 [Fai~
#>  3 doi:~ 1.0.0 ~ 2.00e8 DE-W     18 [Mon~ 2 [8 -~    44 1 [Two~ 1 [Exc~
#>  4 doi:~ 1.0.0 ~ 2.00e8 DE-W     18 [Mon~ 3 [13 ~    55 1 [Two~ 1 [Exc~
#>  5 doi:~ 1.0.0 ~ 2.00e8 DE-W     19 [Tue~ 2 [8 -~    44 1 [Two~ 1 [Exc~
#>  6 doi:~ 1.0.0 ~ 2.00e8 DE-W      7 [Thu~ 2 [8 -~    41 1 [Two~ 1 [Exc~
#>  7 doi:~ 1.0.0 ~ 2.00e8 DE-W     12 [Tue~ 3 [13 ~    29 1 [Two~ 2 [Fai~
#>  8 doi:~ 1.0.0 ~ 2.00e8 DE-W      8 [Fri~ 3 [13 ~    41 1 [Two~ 2 [Fai~
#>  9 doi:~ 1.0.0 ~ 2.00e8 DE-W      7 [Thu~ 2 [8 -~    44 1 [Two~ 1 [Exc~
#> 10 doi:~ 1.0.0 ~ 2.00e8 DE-W      7 [Thu~ 2 [8 -~    44 1 [Two~ 1 [Exc~
#> # ... with 6,569 more rows, and 45 more variables: nuts <chr+lbl>,
#> #   d7 <dbl+lbl>, d8 <dbl+lbl>, d25 <dbl+lbl>, d60 <dbl+lbl>, qa14_5 <dbl+lbl>,
#> #   qa14_3 <dbl+lbl>, qa14_2 <dbl+lbl>, qa14_4 <dbl+lbl>, qa14_1 <dbl+lbl>,
#> #   qa6a_5 <dbl+lbl>, qa6a_10 <dbl+lbl>, qa6b_2 <dbl+lbl>, qa6a_3 <dbl+lbl>,
#> #   qa6a_1 <dbl+lbl>, qa6b_4 <dbl+lbl>, qa6a_8 <dbl+lbl>, qa6a_9 <dbl+lbl>,
#> #   qa6a_4 <dbl+lbl>, qa6a_2 <dbl+lbl>, qa6b_1 <dbl+lbl>, qa6a_6 <dbl+lbl>,
#> #   qa6a_7 <dbl+lbl>, qa6a_11 <dbl+lbl>, qa6b_3 <dbl+lbl>, qd6.1 <dbl+lbl>,
#> #   qd6.2 <dbl+lbl>, qd6.3 <dbl+lbl>, qd6.4 <dbl+lbl>, qd6.5 <dbl+lbl>,
#> #   qd6.6 <dbl+lbl>, qd6.7 <dbl+lbl>, qd6.8 <dbl+lbl>, qd6.9 <dbl+lbl>,
#> #   qd6.10 <dbl+lbl>, qd6.11 <dbl+lbl>, qd6.12 <dbl+lbl>, qd6.13 <dbl+lbl>,
#> #   qd6.14 <dbl+lbl>, qg1b <dbl+lbl>, qg8 <dbl+lbl>, w1 <dbl>, w3 <dbl>,
#> #   wex <dbl>, filename <chr>
#> 
#> [[2]]
#> # A tibble: 5,789 x 48
#>    doi   version uniqid isocntry       p1      p2    p3      p4      p5
#>    <chr> <chr>    <dbl> <chr>    <dbl+lb> <dbl+l> <dbl> <dbl+l> <dbl+l>
#>  1 doi:~ 1.0.0 ~ 1.00e9 DE-W      3 [Mon~ 3 [13 ~    47 1 [Two~ 3 [Ave~
#>  2 doi:~ 1.0.0 ~ 1.00e9 DE-W      8 [Sat~ 2 [8 -~    32 1 [Two~ 1 [Exc~
#>  3 doi:~ 1.0.0 ~ 1.00e9 DE-W      8 [Sat~ 2 [8 -~    46 1 [Two~ 1 [Exc~
#>  4 doi:~ 1.0.0 ~ 1.00e9 DE-W      4 [Tue~ 3 [13 ~    22 1 [Two~ 2 [Fai~
#>  5 doi:~ 1.0.0 ~ 1.00e9 DE-W      5 [Wed~ 2 [8 -~    30 1 [Two~ 2 [Fai~
#>  6 doi:~ 1.0.0 ~ 1.00e9 DE-W      7 [Fri~ 2 [8 -~    29 1 [Two~ 2 [Fai~
#>  7 doi:~ 1.0.0 ~ 1.00e9 DE-W      2 [Sun~ 2 [8 -~    29 1 [Two~ 2 [Fai~
#>  8 doi:~ 1.0.0 ~ 1.00e9 DE-W     10 [Mon~ 2 [8 -~    35 1 [Two~ 2 [Fai~
#>  9 doi:~ 1.0.0 ~ 1.00e9 DE-W      3 [Mon~ 2 [8 -~    32 1 [Two~ 2 [Fai~
#> 10 doi:~ 1.0.0 ~ 1.00e9 DE-W      3 [Mon~ 3 [13 ~    31 1 [Two~ 2 [Fai~
#> # ... with 5,779 more rows, and 39 more variables: nuts <chr+lbl>,
#> #   d7 <dbl+lbl>, d8 <dbl+lbl>, d25 <dbl+lbl>, d60 <dbl+lbl>, qa14_3 <dbl+lbl>,
#> #   qa14_2 <dbl+lbl>, qa14_1 <dbl+lbl>, qa8a_3 <dbl+lbl>, qa8a_9 <dbl+lbl>,
#> #   qa8b_2 <dbl+lbl>, qa8a_1 <dbl+lbl>, qa8a_7 <dbl+lbl>, qa8a_8 <dbl+lbl>,
#> #   qa8a_2 <dbl+lbl>, qa8a_5 <dbl+lbl>, qa8b_1 <dbl+lbl>, qa8a_4 <dbl+lbl>,
#> #   qa8a_6 <dbl+lbl>, qa8a_10 <dbl+lbl>, qa8b_3 <dbl+lbl>, qd7.1 <dbl+lbl>,
#> #   qd7.2 <dbl+lbl>, qd7.3 <dbl+lbl>, qd7.4 <dbl+lbl>, qd7.5 <dbl+lbl>,
#> #   qd7.6 <dbl+lbl>, qd7.7 <dbl+lbl>, qd7.8 <dbl+lbl>, qd7.9 <dbl+lbl>,
#> #   qd7.10 <dbl+lbl>, qd7.11 <dbl+lbl>, qd7.12 <dbl+lbl>, qd7.13 <dbl+lbl>,
#> #   qd7.14 <dbl+lbl>, w1 <dbl>, w3 <dbl>, wex <dbl>, filename <chr>
#> 
#> [[3]]
#> # A tibble: 4,123 x 38
#>    doi   version uniqid isocntry      p1    p3      p4      p5 nuts     
#>    <chr> <chr>    <dbl> <chr>    <dbl+l> <dbl> <dbl+l> <dbl+l> <chr+lbl>
#>  1 doi:~ 2.0.0 ~ 2.01e7 DE-W     8 [Tue~    24 1 [Two~ 1 [Exc~ DE9 [Nie~
#>  2 doi:~ 2.0.0 ~ 2.02e7 DE-W     6 [Sun~    17 1 [Two~ 1 [Exc~ DE2 [Bay~
#>  3 doi:~ 2.0.0 ~ 2.02e7 DE-W     6 [Sun~    18 1 [Two~ 1 [Exc~ DE2 [Bay~
#>  4 doi:~ 2.0.0 ~ 2.02e7 DE-W     8 [Tue~    25 1 [Two~ 1 [Exc~ DE9 [Nie~
#>  5 doi:~ 2.0.0 ~ 2.02e7 DE-W     9 [Wed~    29 1 [Two~ 1 [Exc~ DE9 [Nie~
#>  6 doi:~ 2.0.0 ~ 2.02e7 DE-W     9 [Wed~    29 1 [Two~ 1 [Exc~ DE9 [Nie~
#>  7 doi:~ 2.0.0 ~ 2.02e7 DE-W     4 [Fri~    29 2 [Thr~ 2 [Fai~ DEF [Sch~
#>  8 doi:~ 2.0.0 ~ 2.02e7 DE-W     4 [Fri~    23 1 [Two~ 1 [Exc~ DEF [Sch~
#>  9 doi:~ 2.0.0 ~ 2.02e7 DE-W     4 [Fri~    38 1 [Two~ 2 [Fai~ DEF [Sch~
#> 10 doi:~ 2.0.0 ~ 2.02e7 DE-W     4 [Fri~    27 1 [Two~ 2 [Fai~ DEF [Sch~
#> # ... with 4,113 more rows, and 29 more variables: d7 <dbl+lbl>, d8 <dbl+lbl>,
#> #   d25 <dbl+lbl>, d60 <dbl+lbl>, qa10_3 <dbl+lbl>, qa10_2 <dbl+lbl>,
#> #   qa10_1 <dbl+lbl>, qa7_4 <dbl+lbl>, qa7_2 <dbl+lbl>, qa7_3 <dbl+lbl>,
#> #   qa7_1 <dbl+lbl>, qa7_5 <dbl+lbl>, qd3_1 <dbl+lbl>, qd3_2 <dbl+lbl>,
#> #   qd3_3 <dbl+lbl>, qd3_4 <dbl+lbl>, qd3_5 <dbl+lbl>, qd3_6 <dbl+lbl>,
#> #   qd3_7 <dbl+lbl>, qd3_8 <dbl+lbl>, qd3_9 <dbl+lbl>, qd3_10 <dbl+lbl>,
#> #   qd3_11 <dbl+lbl>, qd3_12 <dbl+lbl>, qd3_13 <dbl+lbl>, qd3_14 <dbl+lbl>,
#> #   w1 <dbl>, w3 <dbl>, filename <chr>
#>